requirements.txt
eventlet==0.33.3
Flask==2.2.2
Flask_Cors==3.0.10
Flask_SocketIO==5.3.6
python_nmap==0.7.1
